🔥 remove stowUpdate
